+++ src/main/java/com/cg/siptracker/repository/NAVRecordRepository.java
@@ -4,7 +4,6 @@ import com.cg.siptracker.model.NAVRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.lang.ScopedValue;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -14,8 +13,8 @@ public interface NAVRecordRepository extends JpaRepository<NAVRecord, Long> {
     boolean existsByFundNameAndDate(String fundName, LocalDate date);
      List<NAVRecord> findByFundNameContainingIgnoreCase(String fundName);
 
-    Optional<NAVRecord> findByFundNameAndNavDate(String fundName, LocalDate date);
+    Optional<NAVRecord> findByFundNameAndDate(String fundName, LocalDate date);
 
-    Optional<NAVRecord> findTopByFundNameOrderByNavDateDesc(String fundName);
+    Optional<NAVRecord> findTopByFundNameOrderByDateDesc(String fundName);
 }
