+++ src/main/java/com/cg/siptracker/repository/UserRepository.java
@@ -8,7 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
     Optional<User> findByEmail(String email);
 //    boolean existsByUsername(String username);
 //    boolean existsByEmail(String email);
