--- src/main/java/com/cg/siptracker/utility/JwtUserDetailsService.java
@@ -1,23 +1,0 @@
-package com.cg.siptracker.utility;
-
-import com.cg.siptracker.model.User;
-import com.cg.siptracker.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class JwtUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("user not found with email"+ email));
-        return user;
-    }
-}
