+++ src/main/resources/application.properties
@@ -1,13 +1,13 @@
 spring.application.name=sip-tracker
 
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/sip_tracker
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
-spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 
 #JMS
 spring.mail.properties.mail.smtp.ssl.protocols=TLSv1.2
